% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crma_query.R
\name{crma_query}
\alias{crma_query}
\title{Query TCRM}
\usage{
crma_query(
  query = NULL,
  debug = FALSE,
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_ID")
)
}
\arguments{
\item{query}{The query, in string format.    Either query or queryString (not used yet) is required, but not both.}

\item{debug}{Outputs the call in the Console for debugging purposes}

\item{consumer_key}{Consumer key given by CRMA App}

\item{consumer_secret}{Consumer secret given by CRMA App}
}
\value{
Dataframe of datasets and metadata
}
\description{
This function will run a query using SOQL
}
