% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crma_oauth.R
\name{crmar_auth}
\alias{crmar_auth}
\title{Description this is the description}
\usage{
crmar_auth(
  baseurl = "https://login.salesforce.com/",
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_SECRET")
)
}
\arguments{
\item{baseurl}{A URL that the Auth should use to authenticate through}

\item{consumer_key}{This is the name of the person}

\item{consumer_secret}{this is the other person}
}
\description{
Description this is the description
}
