% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_version.R
\name{get_dataset_version}
\alias{get_dataset_version}
\title{Get A Dataset's Versions}
\usage{
get_dataset_version(
  datasetId = NULL,
  versionId = NULL,
  debug = FALSE,
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_ID")
)
}
\arguments{
\item{datasetId}{The ID of the dataset}

\item{versionId}{The ID of the dataset version}

\item{debug}{Outputs the call in the Console for debugging purposes}

\item{consumer_key}{Consumer key given by CRMA App}

\item{consumer_secret}{Consumer secret given by CRMA App}
}
\value{
Dataframe of datasets and metadata
}
\description{
This function will pull the version information for a dataset in TCRM
}
