% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasetexportparts.R
\name{get_datasetexportparts}
\alias{get_datasetexportparts}
\title{Get A Dataset Export Part ID}
\usage{
get_datasetexportparts(
  exportid = NULL,
  limit = 100,
  fields = c("Id", "DataFileLength", "DatasetExportId", "CreatedDate", "PartNumber"),
  debug = FALSE,
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_ID")
)
}
\arguments{
\item{exportid}{The export id}

\item{limit}{Requesting Metadata}

\item{fields}{Request specific fields other than the default \verb{Id, CreatedDate, and PartNumber}. See details for more options.}

\item{debug}{Outputs the call in the Console for debugging purposes}

\item{consumer_key}{Consumer key given by CRMA App}

\item{consumer_secret}{Consumer secret given by CRMA App}
}
\value{
Dataframe
}
\description{
This function will pulls a list of available dataset exports
}
\details{
Fields options can include any of the following options:
AllCompressedDataFileLength, CreatedById, CreatedDate, DataFile, DataFileLength,
DatasetExportId, Id, IsDeleted, LastModifiedById, LastModifiedDate, Owner, PartNumber, SystemModstamp
}
