% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasetexports.R
\name{get_datasetexports}
\alias{get_datasetexports}
\title{Get A Dataset Exports}
\usage{
get_datasetexports(
  limit = 100,
  fields = "FIELDS(ALL)",
  debug = FALSE,
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_ID")
)
}
\arguments{
\item{limit}{Requesting Metadata}

\item{fields}{The fields to be included in the response. Default is all fields.}

\item{debug}{Outputs the call in the Console for debugging purposes}

\item{consumer_key}{Consumer key given by CRMA App}

\item{consumer_secret}{Consumer secret given by CRMA App}
}
\value{
Dataframe
}
\description{
This function will pulls a list of available dataset exports
}
