% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasetfile.R
\name{get_datasetfile}
\alias{get_datasetfile}
\title{Get A Dataset Export Files}
\usage{
get_datasetfile(
  partId = NULL,
  debug = FALSE,
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_ID")
)
}
\arguments{
\item{partId}{The ID of the dataset}

\item{debug}{Outputs the call in the Console for debugging purposes}

\item{consumer_key}{Consumer key given by CRMA App}

\item{consumer_secret}{Consumer secret given by CRMA App}
}
\value{
Dataframe of datasets and metadata
}
\description{
This function will pulls a list of files for a dataset from TCRM
}
