% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resources.R
\name{get_resources}
\alias{get_resources}
\title{Get CRMA Resources By Version}
\usage{
get_resources(
  debug = FALSE,
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_ID")
)
}
\arguments{
\item{debug}{Outputs the call in the Console for debugging purposes}

\item{consumer_key}{Consumer key given by CRMA App}

\item{consumer_secret}{Consumer secret given by CRMA App}
}
\value{
Dataframe of datasets and metadata
}
\description{
This function will pull all the available resources
}
