% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_versions.R
\name{get_versions}
\alias{get_versions}
\title{Get Salesforce Versions}
\usage{
get_versions(
  debug = FALSE,
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_ID")
)
}
\arguments{
\item{debug}{Outputs the call in the Console for debugging purposes}

\item{consumer_key}{Consumer key given by CRMA App}

\item{consumer_secret}{Consumer secret given by CRMA App}
}
\value{
Dataframe of versions
}
\description{
This function will pull all the versions available
}
