% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soql_query.R
\name{soql_query}
\alias{soql_query}
\title{Query SOQL}
\usage{
soql_query(
  query = NULL,
  body = FALSE,
  debug = FALSE,
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_ID")
)
}
\arguments{
\item{query}{The query, in string format.    Either query or queryString is required, but not both. (not used yet)}

\item{body}{A query string appending to the body of the api call}

\item{debug}{Outputs the call in the Console for debugging purposes}

\item{consumer_key}{Consumer key given by CRMA App}

\item{consumer_secret}{Consumer secret given by CRMA App}
}
\value{
Dataframe of datasets and metadata
}
\description{
This function will run a query using SOQL
}
